Extract proxy agent support (#117054)
